use int for types.LongType

+++ stdlib/2.7/types.pyi
@@ -13,7 +13,7 @@ TypeType = type
 ObjectType = object
 
 IntType = int
-LongType = long
+LongType = int
 FloatType = float
 BooleanType = bool
 ComplexType = complex
